requirements.txt
streamlit==1.35.0
pandas
scikit-learn
joblib
nltk
numpy
matplotlib
regex
Pillow
